+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.2] - Unreleased
+### Changed
+- `JSONCoStringify`: Added circular reference checks
+- `JSONCoStringify`: Handle `Channel`
+- `pom.xml`: Updated dependency versions
+
 ## [0.1] - 2020-04-19
 ### Added
 - Initial versions of all files